pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-browser"
version = "0.1.0"
description = "A Linux-first agentic browser runner that controls Chromium via Playwright"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Agentic Browser Contributors"}
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Browsers",
]
keywords = ["browser", "automation", "playwright", "llm", "agent"]
dependencies = [
    "playwright>=1.40.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyside6>=6.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
agentic-browser = "agentic_browser.cli:main"

[project.urls]
Homepage = "https://github.com/agentic-browser/agentic-browser"
Repository = "https://github.com/agentic-browser/agentic-browser.git"

[tool.setuptools.packages.find]
where = ["."]
include = ["agentic_browser*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_functions = ["test_*"]
